Separate tools from tests
